requirements.txt
python-docx
langdetect
googletrans==4.0.0-rc1
pymupdf
reportlab